split vtk utilities
